+++ airflow/models/abstractoperator.py
@@ -20,8 +20,6 @@ from __future__ import annotations
 import datetime
 import inspect
 from abc import abstractproperty
-from dataclasses import dataclass
-from datetime import timedelta
 from functools import cached_property
 from typing import TYPE_CHECKING, Any, Callable, ClassVar, Collection, Iterable, Iterator, Sequence
 
@@ -85,24 +83,6 @@ DEFAULT_TASK_EXECUTION_TIMEOUT: datetime.timedelta | None = conf.gettimedelta(
 
 class NotMapped(Exception):
     """Raise if a task is neither mapped nor has any parent mapped groups."""
-
-
-@dataclass
-class StartTriggerArgs:
-    """Arguments required for start task execution from triggerer."""
-
-    trigger_cls: str
-    next_method: str
-    trigger_kwargs: dict[str, Any] | None = None
-    timeout: timedelta | None = None
-
-    def serialize(self):
-        return {
-            "trigger_cls": self.trigger_cls,
-            "trigger_kwargs": self.trigger_kwargs,
-            "next_method": self.next_method,
-            "timeout": self.timeout,
-        }
 
 
 class AbstractOperator(Templater, DAGNode):
